refactor(api): modulariza domínios de morador/funcionário e atualiza plano
- Cria módulos independentes para Morador e Funcionário.
- Isola schemas, controllers e services especializados para o domínio.
- Implementa endpoint e lógica de aprovação para funcionários.
- Atualiza plano de evolução para incluir a nova camada de Model (Etapa 2.6).
- Mantém integridade total da suíte de testes e conformidade com Ruff.
